feat(localstorage): make localstorage for tasks columns and change folder name to utiils
